:zap: Move solo over to bundler
Add minify option to config.
